+++ lib/ubersicht.rb
@@ -1,11 +1,7 @@
-require 'dry-struct'
 require 'faraday'
 
 require 'ubersicht/version'
-require 'ubersicht/ingestion'
-require 'ubersicht/ingestion/build_ingestion_event'
 require 'ubersicht/ingestion/client'
-require 'ubersicht/ingestion/event'
 require 'ubersicht/ingestion/hmac_validator'
 
 module Ubersicht