+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 ## Not released
+
+Breaking changes:
+- [ingestion] changed singnature of from keywords to arguments in `ingest` method
+- [ingestion] removed HMAC key from client constructor
+- [ingestion] removed provider key from client constructor
+
 ## 0.2.0
 
 - [ingestion] added `event.payload.event_id` to prevent events duplication on Ubersicht side
 - [ingestion] added `event.payload.event_id` to event signature
 - [ingestion] added `debug` option to ingestion client. If false then request will be ran in a new thread.
 
-Braking changes:
+Breaking changes:
 
 - [ingestion] replaced `ingest` signature to keywords
 - [ingestion] renamed `event.type` with `event.transaction_type`